Conflicts:
	tools/qdoc3/generator.cpp
	tools/qdoc3/node.cpp

+++ src/multimedia/base/qpaintervideosurface.cpp
@@ -782,10 +782,10 @@ QAbstractVideoSurface::Error QVideoSurfaceArbFpPainter::paint(
         };
         const float v_array[] =
         {
-            target.left()     , target.bottom() + 1,
-            target.right() + 1, target.bottom() + 1,
-            target.left()     , target.top(),
-            target.right() + 1, target.top()
+            float(target.left())     , float(target.bottom() + 1),
+            float(target.right() + 1), float(target.bottom() + 1),
+            float(target.left())     , float(target.top()),
+            float(target.right() + 1), float(target.top())
         };
 
         glEnable(GL_FRAGMENT_PROGRAM_ARB);
@@ -1100,34 +1100,34 @@ QAbstractVideoSurface::Error QVideoSurfaceGlslPainter::paint(
         const GLfloat positionMatrix[4][4] =
         {
             {
-                /*(0,0)*/ wfactor * transform.m11() - transform.m13(),
-                /*(0,1)*/ hfactor * transform.m12() + transform.m13(),
+                /*(0,0)*/ GLfloat(wfactor * transform.m11() - transform.m13()),
+                /*(0,1)*/ GLfloat(hfactor * transform.m12() + transform.m13()),
                 /*(0,2)*/ 0.0,
-                /*(0,3)*/ transform.m13()
+                /*(0,3)*/ GLfloat(transform.m13())
             }, {
-                /*(1,0)*/ wfactor * transform.m21() - transform.m23(),
-                /*(1,1)*/ hfactor * transform.m22() + transform.m23(),
+                /*(1,0)*/ GLfloat(wfactor * transform.m21() - transform.m23()),
+                /*(1,1)*/ GLfloat(hfactor * transform.m22() + transform.m23()),
                 /*(1,2)*/ 0.0,
-                /*(1,3)*/ transform.m23()
+                /*(1,3)*/ GLfloat(transform.m23())
             }, {
                 /*(2,0)*/ 0.0,
                 /*(2,1)*/ 0.0,
                 /*(2,2)*/ -1.0,
                 /*(2,3)*/ 0.0
             }, {
-                /*(3,0)*/ wfactor * transform.dx() - transform.m33(),
-                /*(3,1)*/ hfactor * transform.dy() + transform.m33(),
+                /*(3,0)*/ GLfloat(wfactor * transform.dx() - transform.m33()),
+                /*(3,1)*/ GLfloat(hfactor * transform.dy() + transform.m33()),
                 /*(3,2)*/ 0.0,
-                /*(3,3)*/ transform.m33()
+                /*(3,3)*/ GLfloat(transform.m33())
             }
         };
 
         const GLfloat vertexCoordArray[] =
         {
-            target.left()     , target.bottom() + 1,
-            target.right() + 1, target.bottom() + 1,
-            target.left()     , target.top(),
-            target.right() + 1, target.top()
+            GLfloat(target.left())     , GLfloat(target.bottom() + 1),
+            GLfloat(target.right() + 1), GLfloat(target.bottom() + 1),
+            GLfloat(target.left())     , GLfloat(target.top()),
+            GLfloat(target.right() + 1), GLfloat(target.top())
         };
 
         const GLfloat txLeft = source.left() / m_frameSize.width();
